+++ transport_catalogue.h
@@ -35,7 +35,7 @@ public:
     void addRoute(Route* route);
     void addRoute(const std::string_view name, const std::vector<Stop*>& stops, bool cycled = false);
     void setDistance(Stop* first, Stop* second, int meters);
-    void setDistance(const std::string_view first, const std::string_view second, int meters);
+    void setDistance(const std::string_view first, const std::string_view second, double meters);
     std::optional<Route*> route(const std::string_view name) const;
     std::optional<Stop*> stop(const std::string_view name) const;
     const RouteSet allRoutes() const {return RouteSet(routes_.begin(), routes_.end());}
@@ -43,8 +43,8 @@ public:
     const RouteSet routesViaStop(const std::string_view name) const;
     std::optional<const Route::Info> routeInfo(const std::string_view name) const;
     std::optional<const Stop::Info> stopInfo(const std::string_view name) const;
-    int distanceBetween(Stop* first, Stop* second) const;
-    int distanceBetween(const std::string_view first, const std::string_view second) const;
+    double distanceBetween(Stop* first, Stop* second) const;
+    double distanceBetween(const std::string_view first, const std::string_view second) const;
 
 private:
     std::deque<Stop*> stops_;
@@ -52,7 +52,7 @@ private:
     std::unordered_map<std::string_view, Route*, detail::RouteNumberHasher> nameToBus_;
     std::unordered_map<std::string_view, Stop*, detail::StopNameHasher> nameToStop_;
     std::unordered_map<Stop*, RouteSet> stopToBuses_;
-    std::unordered_map<std::pair<Stop*,Stop*>, int, detail::StopPairHasher> stopsToDistances_;
+    std::unordered_map<std::pair<Stop*,Stop*>, double, detail::StopPairHasher> stopsToDistances_;
 };
 
 } // transport
